Rename sformat_to_mjcf.py
